build: clear all warnings and enable -Werror
Fix the -Wall -Wextra findings at their source:

- fileio.h/core_fileio.c: make read_n_bytes/write_n_bytes/write_padding
  use wlong counts and return wlong, so byte-count comparisons are
  unsigned on both sides
- module.h: mark the CVSID string ST_UNUSED (it is only referenced by
  module initializers) and suppress -Wmissing-field-initializers, since
  the trailing child_args slots are intentionally left for runtime
- use explicit casts at signed/unsigned boundaries and size_t indices
  in the remaining mode/format modules
- silence intentionally-ignored fgets results with (void)!
- meson: default werror=true now that the tree is warning-free

Verified with a clean build under -Werror; smoke and the 33-case
differential suite both pass.

+++ src/format_wv.c
@@ -233,7 +233,7 @@ static bool is_our_file(char *filename) {
 
     remaining_bytes = sizeof(WavpackHeader4) - WV_COMMON_HEADER_SIZE;
     if (fread(wph + WV_COMMON_HEADER_SIZE, 1, remaining_bytes, info->input) !=
-        remaining_bytes) {
+        (size_t)remaining_bytes) {
       fclose(info->input);
       st_free(info);
       return FALSE;
@@ -283,7 +283,7 @@ static bool is_our_file(char *filename) {
 
   remaining_bytes = sizeof(WavpackHeader3) - WV_COMMON_HEADER_SIZE;
   if (fread(wph + WV_COMMON_HEADER_SIZE, 1, remaining_bytes, info->input) !=
-      remaining_bytes) {
+      (size_t)remaining_bytes) {
     fclose(info->input);
     st_free(info);
     return FALSE;
